refactor(purge): mover mapeamento para use case e reorganizar DTO #377
Move o Select/mapeamento de PurgeRecord → PurgeRecordDto para dentro de
GetPurgeRecordsUseCase.ExecuteAsync (retorna IEnumerable<PurgeRecordDto>).
Controller passa a fazer apenas return Ok(result), sem lógica de negócio.
Move PurgeRecordDto de UseCases/Financial/Purge/ para DTOs/Financial/,
alinhando ao padrão do projeto (EligiblePeriodDto).
