+++ src/PersonalFinance.Api/Controllers/V1/Purge/PurgeController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using PersonalFinance.Application.DTOs.Financial;
 using PersonalFinance.Application.UseCases.Financial.Purge;
 using PersonalFinance.Domain.Interfaces.Repositories;
 
@@ -79,18 +80,8 @@ public sealed class PurgeController(
     [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<IActionResult> GetRecords(CancellationToken ct)
     {
-        var records = await _getPurgeRecordsUseCase.ExecuteAsync(CurrentUserId, ct);
-        var dtos = records.Select(r => new PurgeRecordDto
-        {
-            Id           = r.Id,
-            Year         = r.PeriodYear,
-            Month        = r.PeriodMonth,
-            ItemCount    = r.ExpenseCount + r.IncomeCount,
-            TotalIncome  = r.TotalIncome,
-            TotalExpense = r.TotalExpense,
-            PurgedAt     = r.PurgedAt,
-        });
-        return Ok(dtos);
+        var result = await _getPurgeRecordsUseCase.ExecuteAsync(CurrentUserId, ct);
+        return Ok(result);
     }
 
     /// <summary>Remove um registro de expurgo do usuário autenticado.</summary>
